Fix image and clock positioning in HTML file
This commit updates the HTML file to fix the positioning of the image and clock. The image has been resized to be slightly smaller and centered on the page, and the clock has been moved to the top of the page. The CSS has been modified to achieve these changes. This should improve the layout and make it easier to view the image and time simultaneously.